django/requirements.txt
Django>=4.1
djangorestframework
psycopg2==2.9.9
django-cors-headers==4.3.1
